+++ src/main/java/org/example/hydrocore/repository/entity/Produto.java
@@ -12,7 +12,7 @@ public class Produto {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id_produto")
-    private Long idProduto;
+    private Integer idProduto;
 
     @Column(name = "nome_produto")
     private String nomeProduto;
